refactor(guardian-ui): update prettier to use spaces instead of tabs
