+++ guardian-ui/.eslintrc.json
@@ -1,24 +1,23 @@
 {
-	"env": {
-		"browser": true,
-		"es2021": true
-	},
-	"extends": [
-		"eslint:recommended",
-		"plugin:eslint-plugin/recommended",
-		"plugin:@typescript-eslint/recommended",
-		"prettier"
-	],
-	"overrides": [],
-	"parser": "@typescript-eslint/parser",
-	"parserOptions": {
-		"ecmaVersion": "latest",
-		"sourceType": "module"
-	},
-	"plugins": ["eslint-plugin", "@typescript-eslint"],
-	"rules": {
-		"indent": ["warn", "tab"],
-		"quotes": ["error", "single"],
-		"semi": ["error", "always", { "omitLastInOneLineBlock": true }]
-	}
+  "env": {
+    "browser": true,
+    "es2021": true
+  },
+  "extends": [
+    "eslint:recommended",
+    "plugin:eslint-plugin/recommended",
+    "plugin:@typescript-eslint/recommended",
+    "prettier"
+  ],
+  "overrides": [],
+  "parser": "@typescript-eslint/parser",
+  "parserOptions": {
+    "ecmaVersion": "latest",
+    "sourceType": "module"
+  },
+  "plugins": ["eslint-plugin", "@typescript-eslint"],
+  "rules": {
+    "quotes": ["error", "single"],
+    "semi": ["error", "always", { "omitLastInOneLineBlock": true }]
+  }
 }
