+++ guardian-ui/package.json
@@ -1,66 +1,66 @@
 {
-	"name": "guardian-ui",
-	"version": "0.1.0",
-	"private": true,
-	"dependencies": {
-		"@chakra-ui/icons": "^2.0.19",
-		"@chakra-ui/react": "^2.4.5",
-		"@emotion/react": "^11",
-		"@emotion/styled": "^11",
-		"@testing-library/jest-dom": "^5.14.1",
-		"@testing-library/react": "^13.0.0",
-		"@testing-library/user-event": "^13.2.1",
-		"@types/jest": "^27.0.1",
-		"@types/node": "^16.7.13",
-		"@types/react": "^18.0.0",
-		"@types/react-dom": "^18.0.0",
-		"@types/valid-url": "^1.0.3",
-		"framer-motion": "^6",
-		"qrcode.react": "^3.1.0",
-		"react": "^18.2.0",
-		"react-dom": "^18.2.0",
-		"react-icons": "^4.7.1",
-		"react-scripts": "5.0.1",
-		"typescript": "^4.4.2",
-		"valid-url": "^1.0.9",
-		"web-vitals": "^2.1.0"
-	},
-	"scripts": {
-		"start": "react-scripts start",
-		"build": "react-scripts build",
-		"test": "react-scripts test"
-	},
-	"lint-staged": {
-		"**/*.{js,jsx,ts,tsx}": [
-			"npx prettier --write",
-			"npx eslint --fix -c .eslintrc.json"
-		]
-	},
-	"browserslist": {
-		"production": [
-			">0.2%",
-			"not dead",
-			"not op_mini all"
-		],
-		"development": [
-			"last 1 chrome version",
-			"last 1 firefox version",
-			"last 1 safari version"
-		]
-	},
-	"devDependencies": {
-		"@typescript-eslint/eslint-plugin": "^5.48.1",
-		"@typescript-eslint/parser": "^5.48.2",
-		"eslint": "^8.0.1",
-		"eslint-config-prettier": "^8.6.0",
-		"eslint-plugin": "^1.0.1",
-		"eslint-plugin-eslint-plugin": "^5.0.7",
-		"eslint-plugin-import": "^2.25.2",
-		"eslint-plugin-n": "^15.0.0",
-		"eslint-plugin-promise": "^6.0.0",
-		"eslint-plugin-react": "^7.32.0",
-		"lint-staged": "^13.1.0",
-		"prettier": "^2.8.3",
-		"tslint-config-prettier": "^1.18.0"
-	}
+  "name": "guardian-ui",
+  "version": "0.1.0",
+  "private": true,
+  "dependencies": {
+    "@chakra-ui/icons": "^2.0.19",
+    "@chakra-ui/react": "^2.4.5",
+    "@emotion/react": "^11",
+    "@emotion/styled": "^11",
+    "@testing-library/jest-dom": "^5.14.1",
+    "@testing-library/react": "^13.0.0",
+    "@testing-library/user-event": "^13.2.1",
+    "@types/jest": "^27.0.1",
+    "@types/node": "^16.7.13",
+    "@types/react": "^18.0.0",
+    "@types/react-dom": "^18.0.0",
+    "@types/valid-url": "^1.0.3",
+    "framer-motion": "^6",
+    "qrcode.react": "^3.1.0",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-icons": "^4.7.1",
+    "react-scripts": "5.0.1",
+    "typescript": "^4.4.2",
+    "valid-url": "^1.0.9",
+    "web-vitals": "^2.1.0"
+  },
+  "scripts": {
+    "start": "react-scripts start",
+    "build": "react-scripts build",
+    "test": "react-scripts test"
+  },
+  "lint-staged": {
+    "**/*.{js,jsx,ts,tsx}": [
+      "npx prettier --write",
+      "npx eslint --fix -c .eslintrc.json"
+    ]
+  },
+  "browserslist": {
+    "production": [
+      ">0.2%",
+      "not dead",
+      "not op_mini all"
+    ],
+    "development": [
+      "last 1 chrome version",
+      "last 1 firefox version",
+      "last 1 safari version"
+    ]
+  },
+  "devDependencies": {
+    "@typescript-eslint/eslint-plugin": "^5.48.1",
+    "@typescript-eslint/parser": "^5.48.2",
+    "eslint": "^8.0.1",
+    "eslint-config-prettier": "^8.6.0",
+    "eslint-plugin": "^1.0.1",
+    "eslint-plugin-eslint-plugin": "^5.0.7",
+    "eslint-plugin-import": "^2.25.2",
+    "eslint-plugin-n": "^15.0.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.32.0",
+    "lint-staged": "^13.1.0",
+    "prettier": "^2.8.3",
+    "tslint-config-prettier": "^1.18.0"
+  }
 }
