+++ guardian-ui/src/components/ColorModeSwitcher.tsx
@@ -1,30 +1,30 @@
 import * as React from 'react';
 import {
-	useColorMode,
-	useColorModeValue,
-	IconButton,
-	IconButtonProps,
+  useColorMode,
+  useColorModeValue,
+  IconButton,
+  IconButtonProps,
 } from '@chakra-ui/react';
 import { FaMoon, FaSun } from 'react-icons/fa';
 
 type ColorModeSwitcherProps = Omit<IconButtonProps, 'aria-label'>;
 
 export const ColorModeSwitcher: React.FC<ColorModeSwitcherProps> = (props) => {
-	const { toggleColorMode } = useColorMode();
-	const text = useColorModeValue('dark', 'light');
-	const SwitchIcon = useColorModeValue(FaMoon, FaSun);
+  const { toggleColorMode } = useColorMode();
+  const text = useColorModeValue('dark', 'light');
+  const SwitchIcon = useColorModeValue(FaMoon, FaSun);
 
-	return (
-		<IconButton
-			size='md'
-			fontSize='lg'
-			variant='ghost'
-			color='current'
-			marginLeft='2'
-			onClick={toggleColorMode}
-			icon={<SwitchIcon />}
-			aria-label={`Switch to ${text} mode`}
-			{...props}
-		/>
-	);
+  return (
+    <IconButton
+      size='md'
+      fontSize='lg'
+      variant='ghost'
+      color='current'
+      marginLeft='2'
+      onClick={toggleColorMode}
+      icon={<SwitchIcon />}
+      aria-label={`Switch to ${text} mode`}
+      {...props}
+    />
+  );
 };
