+++ guardian-ui/src/index.tsx
@@ -8,8 +8,8 @@ if (!container) throw new Error('Failed to find the root element');
 const root = ReactDOM.createRoot(container);
 
 root.render(
-	<React.StrictMode>
-		<ColorModeScript />
-		<App />
-	</React.StrictMode>
+  <React.StrictMode>
+    <ColorModeScript />
+    <App />
+  </React.StrictMode>
 );
